Combine item create, delete, restore, destroy
